pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tsl-rag"
version = "0.1.0"
description = "RAG system for EU Transport & Logistics compliance"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "openai>=1.35.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "unstructured[pdf]>=0.14.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "pgvector>=0.3.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "tenacity>=8.3.0",
    "tiktoken>=0.7.0",
    "typer>=0.24.1",
    "streamlit>=1.56.0",
    "google-genai>=1.70.0",
]

[project.optional-dependencies]
eval = [
    "ragas>=0.1.14",
    "datasets>=2.20.0",
    "pandas>=2.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "ipykernel>=6.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tsl_rag"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "PTH"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true

[[tool.mypy.overrides]]
module = [
    "rank_bm25.*", "pdfplumber.*", "fitz.*",
    "unstructured.*", "sentence_transformers.*",
    "pgvector.*", "ragas.*", "yaml.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: pure unit tests, no external deps",
    "integration: requires postgres",
    "e2e: full stack",
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Option", "typer.Argument", "fastapi.Depends", "fastapi.Query"]
